[AIRFLOW-3826] Group SageMaker sensors and operators per file
