+++ airflow/contrib/operators/sagemaker_model_operator.py
@@ -17,51 +17,11 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.contrib.hooks.aws_hook import AwsHook
-from airflow.contrib.operators.sagemaker_base_operator import SageMakerBaseOperator
-from airflow.utils.decorators import apply_defaults
-from airflow.exceptions import AirflowException
+import warnings
 
+from airflow.contrib.operators.sagamaker_operator import SageMakerModelOperator # noqa
 
-class SageMakerModelOperator(SageMakerBaseOperator):
-
-    """
-    Create a SageMaker model.
-
-    This operator returns The ARN of the model created in Amazon SageMaker
-
-    :param config: The configuration necessary to create a model.
-
-        For details of the configuration parameter see :py:meth:`SageMaker.Client.create_model`
-    :type config: dict
-    :param aws_conn_id: The AWS connection ID to use.
-    :type aws_conn_id: str
-    """
-
-    @apply_defaults
-    def __init__(self,
-                 config,
-                 *args, **kwargs):
-        super(SageMakerModelOperator, self).__init__(config=config,
-                                                     *args, **kwargs)
-
-        self.config = config
-
-    def expand_role(self):
-        if 'ExecutionRoleArn' in self.config:
-            hook = AwsHook(self.aws_conn_id)
-            self.config['ExecutionRoleArn'] = hook.expand_role(self.config['ExecutionRoleArn'])
-
-    def execute(self, context):
-        self.preprocess_config()
-
-        self.log.info('Creating SageMaker Model %s.', self.config['ModelName'])
-        response = self.hook.create_model(self.config)
-        if response['ResponseMetadata']['HTTPStatusCode'] != 200:
-            raise AirflowException('Sagemaker model creation failed: %s' % response)
-        else:
-            return {
-                'Model': self.hook.describe_model(
-                    self.config['ModelName']
-                )
-            }
+warnings.warn("The operator `SageMakerModelOperator` has been moved to another "
+              "module. You should write "
+              "'from airflow.contrib.operators.segamaker_operator import SageMakerModelOperator`"
+              " to import this operator.", DeprecationWarning)
