+++ docs/code.rst
@@ -250,13 +250,13 @@ Operators
 .. autoclass:: airflow.contrib.operators.sftp_operator.SFTPOperator
 .. autoclass:: airflow.contrib.operators.sftp_to_s3_operator.SFTPToS3Operator
 .. autoclass:: airflow.contrib.operators.ssh_operator.SSHOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_base_operator.SageMakerBaseOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_endpoint_config_operator.SageMakerEndpointConfigOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_endpoint_operator.SageMakerEndpointOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_model_operator.SageMakerModelOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_training_operator.SageMakerTrainingOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_transform_operator.SageMakerTransformOperator
-.. autoclass:: airflow.contrib.operators.sagemaker_tuning_operator.SageMakerTuningOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerBaseOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerEndpointConfigOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerEndpointOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerModelOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerTrainingOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerTransformOperator
+.. autoclass:: airflow.contrib.operators.sagemaker_operator.SageMakerTuningOperator
 .. autoclass:: airflow.contrib.operators.segment_track_event_operator.SegmentTrackEventOperator
 .. autoclass:: airflow.contrib.operators.slack_webhook_operator.SlackWebhookOperator
 .. autoclass:: airflow.contrib.operators.snowflake_operator.SnowflakeOperator
@@ -307,11 +307,11 @@ Sensors
 .. autoclass:: airflow.contrib.sensors.qubole_sensor.QuboleSensor
 .. autoclass:: airflow.contrib.sensors.redis_key_sensor.RedisKeySensor
 .. autoclass:: airflow.contrib.sensors.sftp_sensor.SFTPSensor
-.. autoclass:: airflow.contrib.sensors.sagemaker_base_sensor.SageMakerBaseSensor
-.. autoclass:: airflow.contrib.sensors.sagemaker_endpoint_sensor.SageMakerEndpointSensor
-.. autoclass:: airflow.contrib.sensors.sagemaker_training_sensor.SageMakerTrainingSensor
-.. autoclass:: airflow.contrib.sensors.sagemaker_transform_sensor.SageMakerTransformSensor
-.. autoclass:: airflow.contrib.sensors.sagemaker_tuning_sensor.SageMakerTuningSensor
+.. autoclass:: airflow.contrib.sensors.sagemaker_sensor.SageMakerBaseSensor
+.. autoclass:: airflow.contrib.sensors.sagemaker_sensor.SageMakerEndpointSensor
+.. autoclass:: airflow.contrib.sensors.sagemaker_sensor.SageMakerTrainingSensor
+.. autoclass:: airflow.contrib.sensors.sagemaker_sensor.SageMakerTransformSensor
+.. autoclass:: airflow.contrib.sensors.sagemaker_sensor.SageMakerTuningSensor
 .. autoclass:: airflow.contrib.sensors.wasb_sensor.WasbBlobSensor
 .. autoclass:: airflow.contrib.sensors.wasb_sensor.WasbPrefixSensor
